Modified build_inventory to only rebuild files created during current run. First test of terminating hung autofilet appears to have worked
Previously, build_inventory looked at the whole autofilet.log file. However
when multiple runs were performed, the code crashed when it created warning
arrays used the files in PSUB but then filled them with all of the files
in the log file including files from previous runs. It now only looks at
files in the log file after the most recent run. This could be an issue if
someone runs part way through ORIG, then stops, and then restarts, but the
code will print a warning that the number of files in PSUB does not match
the number of files being rebuild.